lato font, fix open ai fonts, fix cmdinput font
