+++ src/app/overrides.less
@@ -1,15 +1,16 @@
-// Copyright 2023, Command Line Inc.
+// Copyright 2024, Command Line Inc.
 // SPDX-License-Identifier: Apache-2.0
 
 :root {
     // can put dev overrides for testing here
 
-    --label-font-family: Lato;
+    --label-font-family: "Lato", sans-serif;
+
     --sidebar-font-size: 15px;
     --sidebar-font-weight: normal;
     --sidebar-highlight-font-weight: bold;
 
-    --screentabs-font-size: 14px;
+    --screentabs-font-size: 15px;
     --screentabs-font-weight: normal;
     --screentabs-selected-font-weight: bold;
 }
