+++ src/app/root.less
@@ -18,6 +18,8 @@
         "Apple Color Emoji", "Segoe UI Emoji";
     --label-font-family: "Martian Mono", sans-serif;
 
+    --markdown-font-size: 14px;
+
     --screentabs-height: 38px; // magic
     --screentabs-line-height: 21px; // magic
     --screentabs-font-size: 12.5px;
