+++ src/app/workspace/cmdinput/cmdinput.less
@@ -135,6 +135,7 @@
             overflow-wrap: anywhere;
             border-color: transparent;
             border: none;
+            font-family: var(--termfontfamily);
             line-height: var(--termlineheight);
             font-size: var(--termfontsize);
             border-radius: 4px 4px 0 4px; // 0 is for shelltag
